FIONREAD can only return 32-bit size
SCCS-vsn: 7.15

+++ sys/kern/subr_log.c
@@ -30,7 +30,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)subr_log.c	7.14 (Berkeley) 05/20/92
+ *	@(#)subr_log.c	7.15 (Berkeley) 07/07/92
  */
 
 /*
@@ -203,7 +203,7 @@ logioctl(dev, com, data, flag, p)
 		splx(s);
 		if (l < 0)
 			l += MSG_BSIZE;
-		*(off_t *)data = l;
+		*(int *)data = l;
 		break;
 
 	case FIONBIO:
